[FIX]: resolve circular imports in DSC analysis module
- Create new types.py module for shared data classes
- Move DSCPeak and DSCExperiment classes to types.py
- Update imports in core.py and peak_analysis.py
- Maintain same functionality while improving module structure

This change resolves the circular dependency between core.py and
peak_analysis.py by introducing a separate module for shared type
definitions. This follows the principle of separation of concerns
and improves module organization without changing the public API.
