+++ src/pkynetics/technique_analysis/dsc/__init__.py
@@ -1,9 +1,10 @@
 """DSC analysis module for Pkynetics."""
 
 from .baseline import BaselineCorrector
-from .core import DSCAnalyzer, DSCExperiment, DSCPeak
+from .core import DSCAnalyzer
 from .peak_analysis import PeakAnalyzer
 from .thermal_events import ThermalEventDetector
+from .types import DSCExperiment, DSCPeak
 
 __all__ = [
     "DSCAnalyzer",
